Can add items to tiers
- Tier and Item classes have static variable to represent selected
object, if object and tier are selected at the same time, item is
removed from current tier (if exists) and added to chosen tier
- No "untiered" tier exists at present (loaded in items are simply
given xy coords), but this would be as simple as adding another Tier.
Only problem would be, visually, vertical alignment of untiered vs
horizontal of regular tiers (as xy positioning is currently done by
simply adding the width + buffer of each item to the width of the
tier). Regardless, this does need to be fixed anyways, because it
doesn't handle "overflow" from the tier (visually) at all

To-Do:
- Add "untiered" tier (for objects yet to be added to a tier)
- Fix "overflow" from tiers (for when a tier holds too many objects),
likely be scaling tiers vertically
- Tool-tip on object hover to show name (as objects are images)

Potential:

- Add new items when program is running?